+++ src/main/java/com/qa/hwaproject/exceptions/ClubNotFoundException.java
@@ -1,4 +1,4 @@
-package com.qa.hwaproject.com.exceptions;
+package com.qa.hwaproject.exceptions;
 
 import javax.persistence.EntityNotFoundException;
