+++ src/main/java/com/qa/hwaproject/rest/PlayerController.java
@@ -12,7 +12,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 import com.qa.hwaproject.persistence.domain.Player;
+import com.qa.hwaproject.services.PlayerService;
 
 @RestController
 @RequestMapping("/player")
