+++ src/main/java/com/qa/hwaproject/services/ClubService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.qa.hwaproject.com.exceptions.ClubNotFoundException;
+import com.qa.hwaproject.exceptions.ClubNotFoundException;
 import com.qa.hwaproject.persistence.domain.Club;
 import com.qa.hwaproject.persistence.repos.ClubRepo;
